Add files via upload
The work consists of using the R programming language to obtain the sequences of the SARS-CoV-2 variants from the NCBILinks to an external site. or the NCBI virus checkerLinks to an external site. and calculate the length of each sequence. Then, a graph is created where the DNA bases that make up each variant of the virus are compared and the %GC of each of them is calculated. Contrasting sequences for each variant are also created.
To carry out this work, bioinformatics tools and data analysis techniques in R are used. The sequences of the virus variants are downloaded from the NCBILinks database, processed and the length of each one is calculated. . Various R functions are then used to create a graph that allows you to visually compare the DNA bases that make up each variant.
To calculate the %GC of each variant, mathematical formulas are used to count the guanine and cytosine bases present in each sequence and divide them by the total number of bases. To create the antisense sequences, the bases of each sequence are inverted and bases A, C, G, and T are replaced by their corresponding complementary base pairs.
The interpretation of the integrated graphs allows the similarities and differences between the virus variants to be identified. For example, if two variants have a similar DNA base sequence, they may be related to each other. Additionally, %GC can indicate sequence stability and the presence of mutations in the virus. The creation of the antisense sequences allows additional comparisons to be made and evaluation of the structure and function of the proteins that the virus encodes. Overall, this work is important to study and better understand the evolution of SARS-CoV-2 and its impact on public health.